refactor: split mvkroller into common/mvk/plain helper modules
